# perf: improve general performance