The `(async?)` call is resolved through `satisifes?` and is being called several times per interceptor.
On a sufficient long chain, those calls add up and finding if an object implements a given protocol ends up taking some time.

Benchmarking shows around **5ms** on a chain with 100 interceptors:
```
(def ctx {:a 1})
(def sync-ix (into [] (repeat 100 (fn [ctx] (update ctx :a inc)))))
(def async-ix (into [] (repeat 100 (fn [ctx] (d/success-deferred (update ctx :a inc))))))

(comment
  (cc/quick-bench (ix/execute ctx sync-ix)) ;; Execution time mean : 5.580637 ms
  (cc/quick-bench (ixm/execute ctx async-ix))) ;; Execution time mean : 5.534618 ms
```

This PR makes sure sure the `async?` call responds as fast as possible.
Note that `async` is implemented twice for (eg:) deferreds, but the order of requires (`exoscale.interceptor.manifold` depends on  `exoscale.interceptor.protocols`) will ensure that the "correct" implementation wins.

| | before | after | improvement |
|-|----------|--------|--------|
| sync   | ~5.5ms | ~58.5 µs | ~94x faster |
| async | ~5.5ms | ~76.5 µs | ~70x faster |